requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-SQLAlchemy
psycopg2
Flask-WTF
flask-marshmallow


# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gunicorn
